+++ source/Vector.cpp
@@ -14,29 +14,29 @@ namespace cxx {
 template size_t DevicePtr<cuda::Vector<Precision> >::SizeOf();
 template void DevicePtr<cuda::Vector<Precision> >::Construct(
    DevicePtr<Precision> const arr,
-   const int size) const;
+   const size_t size) const;
 template void DevicePtr<cuda::Vector<Precision> >::Construct(
    DevicePtr<Precision> const arr,
-   const int size,
-   const int maxsize) const;
+   const size_t size,
+   const size_t maxsize) const;
 
 template size_t DevicePtr<cuda::Vector<cuda::VPlacedVolume* > >::SizeOf();
 template void DevicePtr<cuda::Vector<cuda::VPlacedVolume* > >::Construct(
    DevicePtr<cuda::VPlacedVolume*> const arr,
-   const int size) const;
+   const size_t size) const;
 template void DevicePtr<cuda::Vector<cuda::VPlacedVolume* > >::Construct(
    DevicePtr<cuda::VPlacedVolume*> const arr,
-   const int size,
-   const int maxsize) const;
+   const size_t size,
+   const size_t maxsize) const;
 
 template size_t DevicePtr<cuda::Vector<cuda::VPlacedVolume const * > >::SizeOf();
 template void DevicePtr<cuda::Vector<cuda::VPlacedVolume const * > >::Construct(
    DevicePtr<cuda::VPlacedVolume const*> const arr,
-   const int size) const;
+   const size_t size) const;
 template void DevicePtr<cuda::Vector<cuda::VPlacedVolume const * > >::Construct(
    DevicePtr<cuda::VPlacedVolume const*> const arr,
-   const int size,
-   const int maxsize) const;
+   const size_t size,
+   const size_t maxsize) const;
 
 } // End cxx namespace
